backend/requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# Vector database
qdrant-client==1.12.1

# Embedding models
sentence-transformers==3.3.1
torch==2.5.1

# Data processing
pandas==2.2.3
numpy==2.2.1

# API utilities
pydantic==2.10.3
pydantic-settings==2.6.1

# CORS support
python-multipart==0.0.18
